Added convenience functionality to store last poses and changed names to symbols
The name of a collision object can now be given as a symbol rather than a string (both are possible). This way, the CRAM-way (specifying object identifiers as symbols) and the MoveIt! string representation stay compatible.

Also, added convenience function changes for leaving out the current pose of an object to be added to the environment. This makes sense when for example a (already known) transported object is detached from the gripper. Its pose is being calculated, but does not need to be set manually again.